feat: make slinky anim 1000% cooler
